Reorganized files, wrote function contracts
